moving button logic out of EventCard
