+++ src/components/EventCard.tsx
@@ -1,23 +1,15 @@
-import { ReactElement, useContext, useEffect, useState } from "react";
+import { ReactElement, useState } from "react";
 import { Link } from "react-router-dom";
 import MeatballMenuSVG from ".././assets/MeatballMenu.svg";
 import PinSVG from ".././assets/Pin.svg";
 import TextIconSVG from ".././assets/TextIcon.svg";
 import ".././styles.css";
-import { LoginContext } from "../App";
 import TriangleError from "../assets/TriangleError.svg";
-import { createAttendanceChange, fetchMember } from "../client/client";
-import {
-  AttendanceChange,
-  ChangeStatus,
-  Event,
-  EventStatus,
-  ListOfButtonClassname,
-} from "../util/Types";
+import { AttendanceChange, Event, EventStatus } from "../util/Types";
+import { AttendanceButton } from "./AttendanceButton";
 import AttendanceChangeModal from "./AttendanceChangeModal";
 import { EventDate } from "./EventDate";
 import EventTag from "./EventTag";
-import Loading from "./Loading";
 import PopUp from "./PopUp";
 
 interface EventCardProps {
@@ -61,34 +53,12 @@ const EventCard = ({
       })
     : [];
 
-  // nuid of the currently loggedIn user
-  const { userID } = useContext(LoginContext);
   const [isRegistered, setIsRegistered] = useState(
     attendanceChange ? false : true
   );
   const [errorType, setErrorType] = useState(0);
   const [isOpen, setIsOpen] = useState(false);
   const [createdAttendanceChange, setCreatedAttendanceChange] = useState({});
-  const [isCreatingAttendance, setIsCreatingAttendance] = useState(false);
-  const [initialAttendanceStatus, setAttendanceStatus] = useState(
-    attendanceChange?.change_status
-  );
-
-  const createButtonText = () => {
-    if (initialAttendanceStatus === ChangeStatus.NOT_REVIEWED) {
-      return ListOfButtonClassname[0];
-    }
-    if (initialAttendanceStatus === ChangeStatus.EXCUSED) {
-      return ListOfButtonClassname[1];
-    }
-    if (initialAttendanceStatus === ChangeStatus.UNEXCUSED) {
-      return ListOfButtonClassname[2];
-    } else {
-      return ListOfButtonClassname[3];
-    }
-  };
-
-  const renderText = createButtonText();
 
   const openModal = () => {
     if (isRegistered) {
@@ -102,31 +72,6 @@ const EventCard = ({
     setIsOpen(false);
     document.body.classList.remove("disable-scrolling");
   };
-
-  useEffect(() => {
-    const makeAttendanceChange = async () => {
-      try {
-        setIsCreatingAttendance(true);
-        //using non-null assertion since it's assumed the user is logged in to make it past the home page
-        const member = await fetchMember(userID!);
-        if (member) {
-          await createAttendanceChange(member.id, id);
-          setIsRegistered(false);
-        }
-        setIsCreatingAttendance(false);
-        // once we successfully created an AttendanceChange its back to pending
-        setAttendanceStatus(ChangeStatus.NOT_REVIEWED);
-      } catch (e) {
-        setErrorType(1);
-        setIsCreatingAttendance(false);
-      }
-    };
-    //on Mount this useEffect starts,
-    //so only want this makeAttendanceChange function to start when we actually have something/ this json is not empty
-    if (!(Object.keys(createdAttendanceChange).length === 0)) {
-      makeAttendanceChange();
-    }
-  }, [createdAttendanceChange, userID, id]);
 
   return (
     <>
@@ -199,12 +144,14 @@ const EventCard = ({
               </button>
             ) : (
               <>
-                <button
-                  onClick={openModal}
-                  className={`${renderText.className}`}
-                >
-                  {isCreatingAttendance ? <Loading /> : renderText.text}
-                </button>
+                <AttendanceButton
+                  openModal={openModal}
+                  setIsRegistered={setIsRegistered}
+                  setErrorType={setErrorType}
+                  eventid={id}
+                  attendanceChange={attendanceChange}
+                  createdAttendanceChange={createdAttendanceChange}
+                />
                 <Link to={`/events/${id}`} state={{ event }}>
                   <button className="button-base-red px-4 my-2 w-32">
                     See More
